NewSentence Based on PatientClient and Tweak NetworkLib
PatientClient now sends a packet to the server everytime it creates a
new sentence, which will then be relayed to group clients. This process
is not fully managed by the server because font size and screen size are
taken into account when new line characters are added, word limit, etc.

Packet class in net.kerfuffle.Utilities.Network now uses CharSequence
instead of String to store data (most notably received data). This is
necessary in order to create proper public constructors for Packets that
are both sent and received, but with different formats respectively.
Helps when a packet has only one String (say a sentence) to send, but
also receives data. This data, previously also being a String, needs to
be something else, CharSequence, in order to avoid a duplicated public
constructor. CharSequence fits the job perfectly because it is read-only
and the data received should not be modified in anyway.